updateOrange

+++ src/common/db/index.ts
@@ -53,12 +53,19 @@ function removeOrange(id: number) {
 }
 
 export
-async function updateOrange(id: string, orange: Partial<IOrange>): Promise<IOrange> {
+async function updateOrange(id: number, orange: {
+  phrase?: string,
+  translation?: string,
+  correctCount?: number,
+  incorrectCount?: number,
+  lastTrainingTime?: number,
+}): Promise<IOrange> {
   const oldOrange = await getOrange(id);
+  if (!oldOrange) throw new Error('target does not exist!');
   return await localForage.setItem(orangeKey(id), {
     ...oldOrange,
     ...orange,
-  } as IOrange);
+  });
 }
 
 export
